feat: add error modal and integrate four modals into one
